Add support for srcset for compliment images
This makes it possible to optionally include alternate widths of
compliment images, which is mainly useful for larger files, e.g. to save
bandwidth on mobile. The alternative versions are specified in the
compliment.imageSrcset object, with the width in pixels as key. Since
the browser doesn’t allow mixing “scale” (2x) and “width“ (240w) srcset
descriptors, and a source without explicit descriptor is equivalent to
“1x”, this means we must also specify the width of the main image (the
src attribute), or else the rest of the srcset will be ignored.

Note that I haven’t been able to test this with the Placeholder.png
here; it seems that Firefox will never request a 240w alternative image
even on very narrow viewports. The feature works fine with a larger,
real compliment image (not committed), though.