Translate to German and add real compliments
Most of the compliments are in German, so it makes sense for the main UI
to be in German too. Consequently, we language-tag English ones, not
German ones; for convenience, introduce an `en` variable that can be
mixed into compliments using ... syntax.

There is one video compliment, and this includes two versions of it (I
hope most browsers support at least one of the two formats). Alternative
versions of images will be added separately.

The order of the compliments has been randomized to obfuscate in what
order they were submitted.

Also, we actually use the term „Gruß“ rather than „Kompliment“, since it
seems more appropriate.